feat: Adding Trivy support
Trivy has been added to replace the role of Grype and Syft.
This has been done as Trivy is the more mature and is considered the defacto tools when working in the world of Kubernetes - IE it appears on the CKS exam where others don't.

Grype and Syft are still cracking tools but since this tool has been developed alongside Baski, which also uses Trivy, it made sense from a toil point of view to stick with the same tool.

There are some minor advantages to using Trivy over Grype/Syft.